rename registers -> registry
