+++ dance/datasets/singlemodality.py
@@ -17,7 +17,7 @@ from dance import logger
 from dance.data import Data
 from dance.datasets.base import BaseDataset
 from dance.metadata.imputation import IMPUTATION_DATASET_TO_FILE
-from dance.registers import register_dataset
+from dance.registry import register_dataset
 from dance.settings import METADIR
 from dance.typing import Dict, List, Optional, Set, Tuple
 from dance.utils.download import download_file, download_unzip
