+++ dance/registry.py
